gaia-magnetics/backend/requirements.txt
# Gaia Geophysics Backend Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
numpy==1.26.2

# AWS
boto3==1.33.6

# Validation and settings
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
pyproj==3.6.1
